chore: logging

+++ server/src/service/token.service.ts
@@ -46,6 +46,11 @@ export class TokenService {
     try {
       log.info('Deleting outdated tokens')
       const tokens = await this.tokenRepo.getOutdatedTokens()
+      if (tokens.length === 0) {
+        log.info('No outdated tokens to delete')
+        return
+      }
+      log.info(`Deleting ${tokens.length} outdated tokens`)
       const deleted = await this.tokenRepo.deleteTokens(tokens)
       log.info(`Deleted ${deleted.length} outdated tokens`)
     } catch (e) {
